requirements.txt
streamlit
openai
python-pptx
pathlib2
pandas
lxml
python-docx
beautifulsoup4
markdown
bcrypt
pymongo
python-dotenv
xlsxwriter
